requirements.txt
streamlit==1.35.0
langchain==0.2.6
langchain-community==0.2.6
pandas==2.2.2
ollama==0.2.1
mysql-connector-python==8.4.0
